requirements.txt
flask
flask-cors
torch
torchvision
Pillow
numpy
matplotlib
scikit-image
werkzeug